+++ src/ApplicationSettings.h
@@ -94,6 +94,7 @@ public:
     DEFINE_SETTING(ShowStatusBar, showStatusBar, bool)
     DEFINE_SETTING(CenterSearchDialog, centerSearchDialog, bool)
     DEFINE_SETTING(Theme, theme, int)
+    DEFINE_SETTING(IconTheme, iconTheme, int)
 
     DEFINE_SETTING(TabsClosable, tabsClosable, bool)
     DEFINE_SETTING(ExitOnLastTabClosed, exitOnLastTabClosed, bool)